Remove old sketches
