+++ visuals/src/main/kotlin/dev/matsem/astral/visuals/di/VisualsModule.kt
@@ -9,19 +9,13 @@ import dev.matsem.astral.visuals.layers.BlobDetectionTerrainLayer
 import dev.matsem.astral.visuals.layers.ConwayLayer
 import dev.matsem.astral.visuals.layers.StarsLayer
 import dev.matsem.astral.visuals.layers.TextOverlayLayer
-import dev.matsem.astral.visuals.sketches.attractor.AttractorSketch
-import dev.matsem.astral.visuals.sketches.boxes.BoxesSketch
-import dev.matsem.astral.visuals.sketches.cubes.CubesSketch
-import dev.matsem.astral.visuals.sketches.fibonaccisphere.FibSphereSketch
-import dev.matsem.astral.visuals.sketches.galaxy.GalaxySketch
-import dev.matsem.astral.visuals.sketches.gameoflife.GameOfLifeSketch
-import dev.matsem.astral.visuals.sketches.oldskool.OldSkoolSketch
-import dev.matsem.astral.visuals.sketches.patterns.PatternsSketch
-import dev.matsem.astral.visuals.sketches.polygonal.PolygonalSketch
-import dev.matsem.astral.visuals.sketches.spikes.SpikesSketch
-import dev.matsem.astral.visuals.sketches.starglitch.StarGlitchSketch
-import dev.matsem.astral.visuals.sketches.terrain.TerrainSketch
-import dev.matsem.astral.visuals.sketches.tunnel.TunnelSketch
+import dev.matsem.astral.visuals.legacy.boxes.BoxesSketch
+import dev.matsem.astral.visuals.legacy.cubes.CubesSketch
+import dev.matsem.astral.visuals.legacy.fibonaccisphere.FibSphereSketch
+import dev.matsem.astral.visuals.legacy.galaxy.GalaxySketch
+import dev.matsem.astral.visuals.legacy.oldskool.OldSkoolSketch
+import dev.matsem.astral.visuals.legacy.patterns.PatternsSketch
+import dev.matsem.astral.visuals.legacy.tunnel.TunnelSketch
 import dev.matsem.astral.visuals.tools.tapper.Tapper
 import dev.matsem.astral.visuals.tools.video.VideoPreparationTool
 import org.koin.dsl.bind
@@ -36,17 +30,11 @@ fun visualsModule(provideSketch: () -> PApplet) = module {
     // BPM
     single { Tapper(get()) }
 
-    factory { PolygonalSketch() }
-    factory { TerrainSketch() }
     factory { FibSphereSketch() }
-    factory { StarGlitchSketch() }
     factory { PatternsSketch() }
     factory { BoxesSketch() }
-    factory { AttractorSketch() }
-    factory { SpikesSketch() }
     factory { CubesSketch() }
     factory { GalaxySketch() }
-    factory { GameOfLifeSketch() }
     factory { OldSkoolSketch() }
     factory { TunnelSketch() }
