+++ visuals/src/main/kotlin/dev/matsem/astral/visuals/legacy/galaxy/GalaxyImage.kt
@@ -1,0 +1,7 @@
+package dev.matsem.astral.visuals.legacy.galaxy
+
+data class GalaxyImage(
+    val path: String,
+    val pixelStep: Int,
+    val threshold: Float
+)
