--- visuals/src/main/kotlin/dev/matsem/astral/visuals/sketches/galaxy/GalaxyImage.kt
@@ -1,7 +1,0 @@
-package dev.matsem.astral.visuals.sketches.galaxy
-
-data class GalaxyImage(
-    val path: String,
-    val pixelStep: Int,
-    val threshold: Float
-)
